將資料庫讀取及邏輯分開，並完成資料庫相關邏輯
