+++ MoneyAPP/ZMoney/App.xaml.cs
@@ -1,7 +1,10 @@
-﻿namespace ZMoney
+﻿using ZMoney.Models;
+namespace ZMoney
 {
     public partial class App : Application
     {
+        public static List<CategoryModel> CachedCategorys { get; set; } = null!;
+        public static List<AccountModel> CachedAccounts { get; set; } = null!;
 
         public App()
         {
