+++ MoneyAPP/ZMoney/MainPage.xaml
@@ -12,11 +12,12 @@
                 HeightRequest="185"
                 Aspect="AspectFit"
                 SemanticProperties.Description="dot net bot in a race car number eight" />
-
-            <Label
-                Text="Hello, World!"
-                Style="{StaticResource Headline}"
-                SemanticProperties.HeadingLevel="Level1" />
+            <Button
+               x:Name="CounterBtn"
+               Text="Click me" 
+               SemanticProperties.Hint="Counts the number of times you click"
+               Clicked="OnCounterClicked"
+               HorizontalOptions="Fill" />
 
             <Label
                 x:Name="LB"
@@ -24,13 +25,6 @@
                 Style="{StaticResource SubHeadline}"
                 SemanticProperties.HeadingLevel="Level2"
                 SemanticProperties.Description="Welcome to dot net Multi platform App U I" />
-
-            <Button
-                x:Name="CounterBtn"
-                Text="Click me" 
-                SemanticProperties.Hint="Counts the number of times you click"
-                Clicked="OnCounterClicked"
-                HorizontalOptions="Fill" />
         </VerticalStackLayout>
     </ScrollView>
